travel-guide-backend/functions/image_processing/requirements.txt
# Image Processing Functions Requirements

# AWS SDK for Python
boto3>=1.26.137

# Python Imaging Library (PIL/Pillow)
# Used for image manipulation, resizing, format conversion
Pillow>=10.0.0

# No additional dependencies needed!
# PIL/Pillow handles all image processing operations